# swagger_output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EGI Media AI Backend API",
    "version": "0.1.0",
    "description": "API contract for the EGI Media AI decision-support service. The service is tenant- and company-scoped, reads editorial articles from egi-media-backend, and exposes only validated AI projections. It never exposes prompts, provider credentials, raw provider output, or CMS write operations."
  },
  "servers": [
    {
      "url": "http://localhost:5003",
      "description": "Local AI backend"
    }
  ],
...
  "paths": {
...
    "/api/v1/internal/pipeline/ingest": {
      "post": {
        "tags": ["Internal Pipeline"],
        "summary": "Start a scoped ingest or backfill run",
        "description": "Internal-only. Does not accept article content from callers. The worker reads CMS using the source contract and performs published/source validation before creating snapshots.",
        "operationId": "startIngestRun",
        "security": [{ "internalServiceAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/IdempotencyKeyHeader" },
          { "$ref": "#/components/parameters/RequestIdHeader" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IngestRunRequest" } } } },
        "responses": {
          "202": { "description": "Ingest run accepted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PipelineRunResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
    "/api/v1/internal/relevance/classify": {
      "post": {
        "tags": ["Internal Relevance"],
        "summary": "Run T02 relevance classification",
        "description": "Runs Nano for one published article and one effective Company Context. `none` is persisted and stops downstream issue formation.",
        "operationId": "classifyArticleRelevance",
        "security": [{ "internalServiceAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["company_id", "article_id"], "properties": { "company_id": { "$ref": "#/components/schemas/Uuid" }, "article_id": { "$ref": "#/components/schemas/Uuid" }, "locale": { "type": "string", "enum": ["id", "en", "uz"], "default": "id" } } } } } },
        "responses": { "200": { "description": "T02 relevance decision" }, "400": { "$ref": "#/components/responses/ValidationError" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "422": { "$ref": "#/components/responses/BusinessRuleError" }, "502": { "$ref": "#/components/responses/ProviderError" } }
      }
    },
    "/api/v1/internal/relevance/rationale": {
      "post": {
        "tags": ["Internal Relevance"],
        "summary": "Run T03 relevance rationale",
        "description": "Runs Nano for an existing T02 decision. The T02 relevance label is immutable and cannot be changed by T03.",
        "operationId": "generateRelevanceRationale",
        "security": [{ "internalServiceAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["company_id", "decision_id"], "properties": { "company_id": { "$ref": "#/components/schemas/Uuid" }, "decision_id": { "$ref": "#/components/schemas/Uuid" } } } } } },
        "responses": { "200": { "description": "T03 rationale and immutable T02 decision" }, "400": { "$ref": "#/components/responses/ValidationError" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "422": { "$ref": "#/components/responses/BusinessRuleError" }, "502": { "$ref": "#/components/responses/ProviderError" } }
      }
    },
    "/api/v1/internal/issues/match": { "post": { "tags": ["Internal Issue Formation"], "summary": "Run T04 issue match", "security": [{ "internalServiceAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["tenant_id", "company_id", "relevance_decision_id"] } } } }, "responses": { "200": { "description": "New/update issue match decision" }, "400": { "$ref": "#/components/responses/ValidationError" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "502": { "$ref": "#/components/responses/ProviderError" } } } },
    "/api/v1/internal/issues/form": { "post": { "tags": ["Internal Issue Formation"], "summary": "Apply T04 issue formation", "description": "Creates a company-scoped issue for new or attaches a development for update. A selesai issue is never silently reopened.", "security": [{ "internalServiceAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["tenant_id", "company_id", "match_decision_id"] } } } }, "responses": { "200": { "description": "Issue mutation result" }, "400": { "$ref": "#/components/responses/ValidationError" }, "403": { "$ref": "#/components/responses/Forbidden" }, "409": { "$ref": "#/components/responses/Conflict" } } } },
    "/api/v1/internal/issues/{issueId}/title": { "post": { "tags": ["Internal Issue Formation"], "summary": "Run T05 issue title", "security": [{ "internalServiceAuth": [] }], "parameters": [{ "name": "issueId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } }, { "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }], "responses": { "200": { "description": "Issue title result" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "409": { "$ref": "#/components/responses/Conflict" }, "502": { "$ref": "#/components/responses/ProviderError" } } } },
    "/api/v1/internal/issues/{issueId}/one-liner": { "post": { "tags": ["Internal Issue Formation"], "summary": "Run T06 issue one-liner", "security": [{ "internalServiceAuth": [] }], "parameters": [{ "name": "issueId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } }, { "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }], "responses": { "200": { "description": "Issue one-liner result" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "409": { "$ref": "#/components/responses/Conflict" }, "502": { "$ref": "#/components/responses/ProviderError" } } } },
...
    "/api/v1/internal/issues/{issueId}/priority": { "post": { "tags": ["Internal Priority"], "summary": "Run T09 priority enum", "description": "Produces only tinggi, sedang, or rendah for a current citation-gated analysis. It does not select Top 5.", "security": [{ "internalServiceAuth": [] }], "parameters": [{ "name": "issueId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } }, { "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }], "responses": { "200": { "description": "Priority decision" }, "403": { "$ref": "#/components/responses/Forbidden" }, "409": { "$ref": "#/components/responses/Conflict" }, "502": { "$ref": "#/components/responses/ProviderError" } } } },
    "/api/v1/internal/issues/{issueId}/priority/reason": { "post": { "tags": ["Internal Priority"], "summary": "Run T10 priority reason", "description": "Produces a grounded reason for the immutable T09 decision. It does not rank or select Top 5.", "security": [{ "internalServiceAuth": [] }], "parameters": [{ "name": "issueId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } }, { "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }], "responses": { "200": { "description": "Priority reason" }, "403": { "$ref": "#/components/responses/Forbidden" }, "409": { "$ref": "#/components/responses/Conflict" }, "502": { "$ref": "#/components/responses/ProviderError" } } } },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "End-user identity. Tenant/company scope is derived and verified server-side." },
      "internalServiceAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "Service JWT", "description": "Internal service or authorized operator identity; never available to frontend." }
    },
    "parameters": {
...
      "IdempotencyKeyHeader": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 16, "maxLength": 255 } },
...
      "RequestIdHeader": { "name": "X-Request-Id", "in": "header", "required": false, "schema": { "type": "string", "maxLength": 100 } }
    },
    "responses": {
      "Unauthorized": { "description": "Authentication missing or invalid", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "Forbidden": { "description": "Caller lacks required tenant/company/role permission", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
...
      "Conflict": { "description": "Idempotency key, version, or state conflict", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "ValidationError": { "description": "Request schema validation failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "BusinessRuleError": { "description": "Valid request cannot proceed because of pipeline/state/delivery business rules", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "ProviderError": { "description": "External AI/provider request failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
...
    },
    "schemas": {
      "Uuid": { "type": "string", "format": "uuid" },
      "Timestamp": { "type": "string", "format": "date-time" },
...
      "PipelineStageName": { "type": "string", "enum": ["ingest", "relevance", "issue_match", "issue_title", "issue_oneliner", "analysis", "claim_label", "priority", "alert_eligibility", "direct_blurb", "report_narrative", "delivery"] },
      "PipelineStageState": { "type": "string", "enum": ["queued", "running", "succeeded", "failed", "retrying", "dead_letter", "waiting_review", "insufficient_data", "stale", "cancelled"] },
      "ErrorCode": { "type": "string", "enum": ["VALIDATION_ERROR", "UNAUTHORIZED", "FORBIDDEN", "NOT_FOUND", "VERSION_CONFLICT", "IDEMPOTENCY_CONFLICT", "BUSINESS_RULE_FAILED", "AI_OUTPUT_INVALID", "SOURCE_VALIDATION_FAILED", "PIPELINE_PENDING", "INTERNAL_ERROR"] },
      "ApiMeta": {
        "type": "object",
        "required": ["request_id"],
        "properties": {
          "request_id": { "type": "string" },
          "trace_id": { "type": "string" },
          "page": { "type": "integer", "minimum": 1 },
          "limit": { "type": "integer", "minimum": 1 },
          "total": { "type": "integer", "minimum": 0 },
          "total_pages": { "type": "integer", "minimum": 0 }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "error", "meta"],
        "properties": {
          "success": { "type": "boolean", "enum": [false] },
          "error": { "type": "object", "required": ["code", "message"], "properties": { "code": { "$ref": "#/components/schemas/ErrorCode" }, "message": { "type": "string" }, "fields": { "type": "array", "items": { "type": "object", "properties": { "field": { "type": "string" }, "code": { "type": "string" } } } } } },
          "meta": { "$ref": "#/components/schemas/ApiMeta" }
        }
      },
...
      "PipelineStage": {
        "type": "object",
        "required": ["name", "state", "updated_at"],
        "properties": {
          "name": { "$ref": "#/components/schemas/PipelineStageName" },
          "state": { "$ref": "#/components/schemas/PipelineStageState" },
          "updated_at": { "$ref": "#/components/schemas/Timestamp" },
          "failure_code": { "type": "string", "description": "Safe classification only; no raw provider/internal error." },
          "next_retry_at": { "$ref": "#/components/schemas/Timestamp" }
        }
      },
...
      "IngestRunRequest": { "type": "object", "required": ["mode", "locale"], "additionalProperties": false, "properties": { "mode": { "type": "string", "enum": ["poll", "article"] }, "article_id": { "$ref": "#/components/schemas/Uuid" }, "limit": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 }, "locale": { "type": "string", "enum": ["id", "en", "uz"] } } },
...
      "PipelineRun": { "type": "object", "required": ["id", "trigger", "state", "stages", "created_at", "trace_id"], "properties": { "id": { "$ref": "#/components/schemas/Uuid" }, "trigger": { "type": "string", "enum": ["scheduled", "manual", "backfill", "replay", "context_draft"] }, "state": { "$ref": "#/components/schemas/PipelineStageState" }, "tenant_id": { "$ref": "#/components/schemas/Uuid" }, "company_id": { "$ref": "#/components/schemas/Uuid" }, "stages": { "type": "array", "items": { "$ref": "#/components/schemas/PipelineStage" } }, "created_at": { "$ref": "#/components/schemas/Timestamp" }, "completed_at": { "$ref": "#/components/schemas/Timestamp" }, "trace_id": { "type": "string" } } },
...
      "PipelineRunResponse": { "type": "object", "required": ["success", "data", "meta"], "properties": { "success": { "type": "boolean", "enum": [true] }, "data": { "$ref": "#/components/schemas/PipelineRun" }, "meta": { "$ref": "#/components/schemas/ApiMeta" } } },
...
    }
  }
}